Updated adapter

+++ internal/domain/entity/user.go
@@ -33,7 +33,7 @@ type Email struct {
 
 type LoginResult struct {
 	AccessToken  *string `json:"access_token"`
-	ExpiresIn    *int64  `json:"expires_in"`
+	ExpiresIn    *int32  `json:"expires_in"`
 	IdToken      *string `json:"id_token"`
 	RefreshToken *string `json:"refresh_token"`
 	TokenType    *string `json:"token_type"`
